requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.8.2
cryptography==45.0.2
docutils==0.21.2
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.10
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rsa==4.9.1
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tle-faster-api==0.1.0
twine==6.1.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
